requirements.txt
telethon==1.28.5
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
flask